Add return types to templates hook
